requirements.txt
aniso8601==4.0.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autoenv==1.0.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-JWT-Extended==3.13.1
flask-reqparse==2.2.1
Flask-RESTful==0.3.6
gunicorn==19.9.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
pep8==1.7.1
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
PyJWT==1.7.0
pylint==2.2.2
pytest==4.0.1
pytest-cov==2.6.0
python-dotenv==0.9.1
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
